fix: exported type move to types.ts
